Trying to add links in descriptions
Attempting to inject HTML with the description parameter. This shouldn't work, right?

If you're wondering why I'm testing on *main* and on my live deployment, well, uhh good point! Currently not on a machine set up to build this site and it's suuuuper low traffic, so this isn't exactly risky for my 1-person toy site.